// src/types/academic.ts
export interface Curso {
    id: number;
    nombre: string;
    nivel: string; // 'PRIMARIA' | 'SECUNDARIA'
    materias: number[];
}